hook/requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
gunicorn==20.1.0
python-dotenv==1.0.0
#ssh-python
paramiko  # ssh